requirements.txt
# Web framework
fastapi==0.111.1
uvicorn==0.23.2

# Vector database
qdrant-client==1.15.1

# Embeddings (local, free model)
sentence-transformers==2.2.2
torch>=2.0.0
transformers>=4.40.0

# File parsing
pdfminer.six==20221105
python-docx==1.2.0
pandas==2.1.0

# Data handling
numpy>=1.26.0

# Environment variables
python-dotenv==1.0.1





